Clean up of the project to remove all the files that are not needed for the project to run. this includes output test files from years ago.
This is the first step in a long process to clean up the project and make it more manageable as well as to update the project to use the latest version of Java.

I will be working on this project for the next few weeks to get it up to date and then port it over to a flutter project to run natively on Android and windows.