Reorgnaized the project structure to follow Maven conventions. The project is now structured as a Maven project with the following layout: Poker-Basic ├── pom.xml ├── src │   ├── main │   │   ├── java │   │   │   └── com │   │   │       └── pokermon │   │   │           ├── Main.java │   │   │           └── NewJFrame.java │   │   └── resources │   │       └── Cards │   │           └── TET │   │               ├── Ace of Clubs.jpg │   │               ├── Ace of Diamonds.jpg │   │               ├── Ace of Hearts (2).jpg │   │               ├── Ace of Hearts.jpg │   │               ├── Ace of Spades.jpg  .. Etc...
The proejct can be run successfully but build testing has not been done yet.
After successful build and testing, the project will be pushed to the main branch and refactored to utilize flutter.

It would also be a great idea to refactor the code to use better naming convertions and structures.